requirements.txt
ansible==8.6.1

# molecule
pytest==7.2.1
pytest-testinfra==7.0.0
molecule==6.0.2
molecule-qemu==0.5.7
